+++ api/v1alpha1/common_types.go
@@ -139,6 +139,9 @@ const (
 
 var log = ctrl.Log.WithName("validate-webhook")
 
+// +kubebuilder:object:root=false
+// +k8s:deepcopy-gen=false
+// +k8s:gen-deepcopy=false
 type StatefulObject interface {
 	runtime.Object
 	GetStatus() *ChaosStatus
